requirements.txt
torch==2.2.2+cu118
torchvision==0.17.2+cu118
scikit-image==0.23.1
omegaconf==2.3.0
matplotlib==3.8.0
opencv-python==4.8.0.76
dotmap==1.3.30
kornia==0.7.0
hydra-core==1.3.2
plyfile==1.1
dotdict==0.1
tqdm==4.65.2
tensorboardX==2.6.2.2
